Created main menu and refined gameplay
V 11.07.17

Added Main Menu:
Has 3 buttons for now, play game, options, and how to play. Only play
game works for now, however options shouldn't be too trivial.
How to play will be a short GIF/ MP4 of someone playing with
instructions on screen

Added Multiple rounds:
Rounds play sequentially with a different image each time. No
duplicates.

Added Game class:
I thought this might be needed but i don't think it'll be used.

Added Version:
At the bottom of the left corner there is a version number in unity.
This contains the current version of the game so that its easier to
debug/ know the issues.

Map color changed:
Filter applied to the map to make it look more vibrant and more fun.