V. 03.19.18
Added a line that connects the guess and real locations.
At the end of the game all the lines and previous guesses are revealed
Added all the locations
Added a touchscreen zoom:
It works well on small screens but is not needed on the large 4k
displays, therefore it is disabled.